+++ ORBIT/phases/design/mooring_system_design.py
@@ -14,6 +14,14 @@ from scipy.interpolate import interp1d
 
 from ORBIT.phases.design import DesignPhase
 
+"""
+[1] Maness et al. 2017, NREL Offshore Balance-of-System Model.
+https://www.nrel.gov/docs/fy17osti/66874.pdf
+
+[2] Cooperman et al. (2022), Assessment of Offshore Wind Energy Leasing Areas
+for Humboldt and Morry Bay. https://www.nrel.gov/docs/fy22osti/82341.pdf
+"""
+
 
 class MooringSystemDesign(DesignPhase):
     """Mooring System and Anchor Design."""
@@ -68,8 +76,7 @@ class MooringSystemDesign(DesignPhase):
         self.anchor_type = self._design.get("anchor_type", "Suction Pile")
         self.mooring_type = self._design.get("mooring_type", "Catenary")
 
-        # Semi-Taut mooring system design parameters based on depth
-        # Cooperman et al. (2022), https://www.nrel.gov/docs/fy22osti/82341.pdf
+        # Semi-Taut mooring system design parameters based on depth [2].
         self._semitaut_params = {
             "depths": [500.0, 750.0, 1000.0, 1250.0, 1500.0],
             "rope_lengths": [478.41, 830.34, 1229.98, 1183.93, 1079.62],
@@ -108,26 +115,32 @@ class MooringSystemDesign(DesignPhase):
         tr = self.config["turbine"]["turbine_rating"]
         fit = -0.0004 * (tr**2) + 0.0132 * tr + 0.0536
 
+        _key = "mooring_line_cost_rate"
+
+        mooring_line_cost_rate = self._design.get(
+            _key,
+            self.get_default_cost(
+                "mooring_system_design",
+                _key,
+            ),
+        )
+        if isinstance(mooring_line_cost_rate, (int, float)):
+            mooring_line_cost_rate = [mooring_line_cost_rate] * 3
+
         if fit <= 0.09:
             self.line_diam = 0.09
             self.line_mass_per_m = 0.161
-            self.line_cost_rate = self._design.get(
-                "mooring_line_cost_rate", 399.0
-            )
+            self.line_cost_rate = mooring_line_cost_rate[0]
 
         elif fit <= 0.12:
             self.line_diam = 0.12
             self.line_mass_per_m = 0.288
-            self.line_cost_rate = self._design.get(
-                "mooring_line_cost_rate", 721.0
-            )
+            self.line_cost_rate = mooring_line_cost_rate[1]
 
         else:
             self.line_diam = 0.15
             self.line_mass_per_m = 0.450
-            self.line_cost_rate = self._design.get(
-                "mooring_line_cost_rate", 1088.0
-            )
+            self.line_cost_rate = mooring_line_cost_rate[2]
 
     def calculate_breaking_load(self):
         """Returns the mooring line breaking load."""
@@ -214,8 +227,8 @@ class MooringSystemDesign(DesignPhase):
         """
         Returns the mass and cost of anchors.
 
-        TODO: Anchor masses are rough estimates based on initial literature
-        review. Should be revised when this module is overhauled in the future.
+        TODO: Anchor masses are rough estimates based on [1]. Should be
+        revised when this module is overhauled in the future.
         TODO: Mooring types for Catenary, TLP, SemiTaut will likely have
         different anchors.
         """