+++ ORBIT/phases/design/semi_submersible_design.py
@@ -1,4 +1,4 @@
-"""Provides the `SemiSubmersibleDesign` class (from OffshoreBOS)."""
+"""Provides the `SemiSubmersibleDesign` class."""
 
 __author__ = "Jake Nunemaker"
 __copyright__ = "Copyright 2020, National Renewable Energy Laboratory"
@@ -8,6 +8,11 @@ __email__ = "jake.nunemaker@nrel.gov"
 import numpy as np
 
 from ORBIT.phases.design import DesignPhase
+
+"""
+[1] Maness et al. 2017, NREL Offshore Balance-of-System Model.
+https://www.nrel.gov/docs/fy17osti/66874.pdf
+"""
 
 
 class SemiSubmersibleDesign(DesignPhase):
@@ -62,90 +67,98 @@ class SemiSubmersibleDesign(DesignPhase):
 
     @property
     def stiffened_column_mass(self):
-        """
-        Calculates the mass of the stiffened column for a single
-        semi-submersible in tonnes. From original OffshoreBOS model.
+        """Calculates the mass of the stiffened column for a single
+        semi-submersible in tonnes [1].
         """
 
         rating = self.config["turbine"]["turbine_rating"]
+
         mass = -0.9581 * rating**2 + 40.89 * rating + 802.09
 
         return mass
 
     @property
     def stiffened_column_cost(self):
-        """
-        Calculates the cost of the stiffened column for a single
-        semi-submersible. From original OffshoreBOS model.
+        """Calculates the cost of the stiffened column for a single
+        semi-submersible [1].
         """
 
-        cr = self._design.get("stiffened_column_CR", 3120)
+        _key = "stiffened_column_CR"
+        cr = self._design.get(
+            _key, self.get_default_cost("semisubmersible_design", _key)
+        )
         return self.stiffened_column_mass * cr
 
     @property
     def truss_mass(self):
-        """
-        Calculates the truss mass for a single semi-submersible in tonnes. From
-        original OffshoreBOS model.
+        """Calculates the truss mass for a single semi-submersible in tonnes
+        [1].
         """
 
         rating = self.config["turbine"]["turbine_rating"]
+
         mass = 2.7894 * rating**2 + 15.591 * rating + 266.03
 
         return mass
 
     @property
     def truss_cost(self):
-        """
-        Calculates the cost of the truss for a signle semi-submerisble. From
-        original OffshoreBOS model.
+        """Calculates the cost of the truss for a signle semi-submerisble
+        [1].
         """
 
-        cr = self._design.get("truss_CR", 6250)
+        _key = "truss_CR"
+        cr = self._design.get(
+            _key, self.get_default_cost("semisubmersible_design", _key)
+        )
         return self.truss_mass * cr
 
     @property
     def heave_plate_mass(self):
-        """
-        Calculates the heave plate mass for a single semi-submersible in
-        tonnes. Source: original OffshoreBOS model.
+        """Calculates the heave plate mass for a single semi-submersible
+        in tonnes [1].
         """
 
         rating = self.config["turbine"]["turbine_rating"]
+
         mass = -0.4397 * rating**2 + 21.545 * rating + 177.42
 
         return mass
 
     @property
     def heave_plate_cost(self):
-        """
-        Calculates the heave plate cost for a single semi-submersible. From
-        original OffshoreBOS model.
+        """Calculates the heave plate cost for a single semi-submersible
+        [1].
         """
 
-        cr = self._design.get("heave_plate_CR", 6250)
+        _key = "heave_plate_CR"
+        cr = self._design.get(
+            _key, self.get_default_cost("semisubmersible_design", _key)
+        )
         return self.heave_plate_mass * cr
 
     @property
     def secondary_steel_mass(self):
-        """
-        Calculates the mass of the required secondary steel for a single
-        semi-submersible. From original OffshoreBOS model.
+        """Calculates the mass of the required secondary steel for a single
+        semi-submersible [1].
         """
 
         rating = self.config["turbine"]["turbine_rating"]
+
         mass = -0.153 * rating**2 + 6.54 * rating + 128.34
 
         return mass
 
     @property
     def secondary_steel_cost(self):
-        """
-        Calculates the cost of the required secondary steel for a single
-        semi-submersible. For original OffshoreBOS model.
+        """Calculates the cost of the required secondary steel for a single
+        semi-submersible [1].
         """
 
-        cr = self._design.get("secondary_steel_CR", 7250)
+        _key = "secondary_steel_CR"
+        cr = self._design.get(
+            _key, self.get_default_cost("semisubmersible_design", _key)
+        )
         return self.secondary_steel_mass * cr
 
     @property